# scripts/tmp/openapi.json
{"openapi":"3.1.0","info":{"title":"DirectorOS Actions API (v0.3.1a)","version":"0.3.1a"},"paths":{"/health":{"get":{"summary":"Health","operationId":"getHealth","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/privacy":{"get":{"summary":"Privacy","operationId":"getPrivacy","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/derive-concepts":{"post":{"summary":"Derive Concepts","operationId":"deriveConcepts","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeriveConceptsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeriveConceptsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/compose-stills":{"post":{"summary":"Compose Stills","operationId":"composeStills","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComposeStillsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComposeStillsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/expand-scene":{"post":{"summary":"Expand Scene","operationId":"expandScene","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpandSceneRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpandSceneResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/qa-validate":{"post":{"summary":"Qa Validate","operationId":"qaValidate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QAValidateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QAValidateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"AlbumInfo":{"properties":{"title":{"type":"string","title":"Title"},"style":{"type":"string","title":"Style"},"lyrics":{"type":"string","title":"Lyrics"}},"type":"object","required":["title","style","lyrics"],"title":"AlbumInfo"},"ComposeStillsRequest":{"properties":{"selection":{"$ref":"#/components/schemas/Concept"},"count":{"type":"integer","maximum":6.0,"minimum":1.0,"title":"Count","default":3}},"type":"object","required":["selection"],"title":"ComposeStillsRequest"},"ComposeStillsResponse":{"properties":{"stills":{"items":{"$ref":"#/components/schemas/MJPrompt"},"type":"array","title":"Stills"}},"type":"object","required":["stills"],"title":"ComposeStillsResponse"},"Concept":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"logline":{"type":"string","title":"Logline"},"style":{"type":"string","title":"Style"},"cast":{"items":{"type":"string"},"type":"array","title":"Cast"},"anchors":{"additionalProperties":{"items":{"type":"string"},"type":"array"},"type":"object","title":"Anchors"},"storyline":{"type":"string","title":"Storyline"}},"type":"object","required":["id","title","logline","style","storyline"],"title":"Concept"},"Controls":{"properties":{"variants":{"type":"integer","maximum":6.0,"minimum":1.0,"title":"Variants","default":3},"palette_override":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Palette Override"},"lighting_override":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Lighting Override"}},"type":"object","title":"Controls"},"DeriveConceptsRequest":{"properties":{"album_info":{"$ref":"#/components/schemas/AlbumInfo"},"controls":{"anyOf":[{"$ref":"#/components/schemas/Controls"},{"type":"null"}]}},"type":"object","required":["album_info"],"title":"DeriveConceptsRequest"},"DeriveConceptsResponse":{"properties":{"concepts":{"items":{"$ref":"#/components/schemas/Concept"},"type":"array","title":"Concepts"}},"type":"object","required":["concepts"],"title":"DeriveConceptsResponse"},"ExpandSceneRequest":{"properties":{"brief":{"type":"string","title":"Brief"},"selection":{"$ref":"#/components/schemas/Concept"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"duration_sec":{"type":"number","title":"Duration Sec","default":3.0},"beats":{"type":"integer","maximum":12.0,"minimum":1.0,"title":"Beats","default":5}},"type":"object","required":["brief","selection"],"title":"ExpandSceneRequest"},"ExpandSceneResponse":{"properties":{"scene":{"$ref":"#/components/schemas/SceneDraft"}},"type":"object","required":["scene"],"title":"ExpandSceneResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MJPrompt":{"properties":{"id":{"type":"string","title":"Id"},"prompt":{"type":"string","title":"Prompt"},"ar":{"type":"string","title":"Ar","default":"16:9"},"notes":{"additionalProperties":true,"type":"object","title":"Notes"}},"type":"object","required":["id","prompt"],"title":"MJPrompt"},"QAReport":{"properties":{"story_match_score":{"type":"number","title":"Story Match Score"},"coverage_score":{"type":"number","title":"Coverage Score"},"conflicts":{"items":{"type":"string"},"type":"array","title":"Conflicts"},"severity":{"type":"string","enum":["info","warn","fail"],"title":"Severity"}},"type":"object","required":["story_match_score","coverage_score","conflicts","severity"],"title":"QAReport"},"QAValidateRequest":{"properties":{"scene":{"$ref":"#/components/schemas/SceneDraft"},"engine":{"type":"string","enum":["SORA","VEO"],"title":"Engine","default":"SORA"}},"type":"object","required":["scene"],"title":"QAValidateRequest"},"QAValidateResponse":{"properties":{"engine_render":{"type":"string","title":"Engine Render"},"report":{"$ref":"#/components/schemas/QAReport"}},"type":"object","required":["engine_render","report"],"title":"QAValidateResponse"},"SceneDraft":{"properties":{"st":{"additionalProperties":true,"type":"object","title":"St"},"ds":{"additionalProperties":true,"type":"object","title":"Ds"},"dw":{"additionalProperties":true,"type":"object","title":"Dw"},"timeline":{"items":{"$ref":"#/components/schemas/TimelineBeat"},"type":"array","title":"Timeline"},"drafts":{"additionalProperties":{"type":"string"},"type":"object","title":"Drafts"}},"type":"object","required":["st","ds","dw","timeline","drafts"],"title":"SceneDraft"},"TimelineBeat":{"properties":{"t":{"type":"number","title":"T"},"beat":{"type":"string","title":"Beat"},"intensity":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Intensity"}},"type":"object","required":["t","beat","intensity"],"title":"TimelineBeat"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}},"servers":[{"url":"https://avg-production.up.railway.app/","description":"Production (Railway)"}]}